Enum volta::command::list::Package

enum Package {
    Default {
        details: PackageDetails,
        node: Version,
        tools: Vec<String>,
    },
    Project {
        details: PackageDetails,
        node: Version,
        tools: Vec<String>,
        path: PathBuf,
    },
    Fetched(PackageDetails),
}

Variants

Default

Fields of Default

details: PackageDetailsnode: Version

The version of Node the package is installed against.

tools: Vec<String>

The names of the tools associated with the package.

Project

Fields of Project

details: PackageDetailsnode: Version

The version of Node the package is installed against.

tools: Vec<String>

The names of the tools associated with the package.

path: PathBuf

Fetched(PackageDetails)

Methods

impl Package

fn new(config: &PackageConfig, source: &Source) -> Package

fn from_inventory_and_project(
inventory: &Inventory,
project: Option<&Project>
) -> Vec<Package>

fn source(name: &str, version: &Version, project: Option<&Project>) -> Source