Struct volta::command::run::Run

pub(crate) struct Run {
    node: Option<String>,
    npm: Option<String>,
    bundled_npm: bool,
    pnpm: Option<String>,
    no_pnpm: bool,
    yarn: Option<String>,
    no_yarn: bool,
    envs: Vec<String>,
    command_and_args: Vec<OsString>,
}

Fields
node: Option<String>

Set the custom Node version

npm: Option<String>

Set the custom npm version

bundled_npm: bool

Forces npm to be the version bundled with Node

pnpm: Option<String>

Set the custon pnpm version

no_pnpm: bool

Disables pnpm

yarn: Option<String>

Set the custom Yarn version

no_yarn: bool

Disables Yarn

envs: Vec<String>

Set an environment variable (can be used multiple times)

command_and_args: Vec<OsString>

The command to run, along with any arguments

Implementations

impl Run

fn parse_platform(&self, session: &mut Session) -> Fallible<CliPlatform>

Builds a CliPlatform from the provided cli options

Will resolve a semver / tag version if necessary

fn parse_envs(&self) -> HashMap<&str, &str>

Convert the environment variable settings passed to the command line into a map

We ignore any setting that doesn’t have a value associated with it We also ignore the PATH environment variable as that is set when running a command